pedidos funcionais
detalhe: carrinho não é excluído do banco ao criar pedido e sim criado novo registro

+++ src/main/java/dao/ItemCarrinhoDAO.java
@@ -30,7 +30,7 @@ public class ItemCarrinhoDAO {
         return em.createQuery("SELECT i FROM ItemCarrinho i", ItemCarrinho.class).getResultList();
     }
 
-    public List<ItemCarrinho> buscarPorCarrinho(Long idCarrinho) {
+    public List<ItemCarrinho> buscarPorCarrinho(int idCarrinho) {
         return em.createQuery("SELECT i FROM ItemCarrinho i WHERE i.carrinho.id = :idCarrinho", ItemCarrinho.class)
                 .setParameter("idCarrinho", idCarrinho)
                 .getResultList();
